// added add_note_walgreen.rs + changed u16 to u32 in routes

+++ src/axum_routes/routes/ugo_vape/orders_routes/add_note_to_order/add_note_to_order_structs.rs
@@ -7,6 +7,6 @@ pub struct AddNoteToOrder {
 }
 
 pub struct InsertStruct {
-    pub id : u16,
+    pub id : u32,
     pub note : String
 }
